Refactored headers. There were repitition in names: sunscriptions.h and subscription.h. I have moved all services headers into a corresponding folder. All high level wrappers are in the top sirectory.Fixed stopping of server: there was a deadlock on DbMutex at address space. I have removed locking it from destructor because object is invalid state and at this moment cannot process requests.
