+++ include/opc/ua/server/addons/standard_namespace.h
@@ -11,7 +11,7 @@
 #pragma once
 
 #include <opc/common/addons_core/addon.h>
-#include <opc/ua/node_management.h>
+#include <opc/ua/services/node_management.h>
 
 namespace OpcUa
 {
