+++ include/opc/ua/server/address_space.h
@@ -10,11 +10,11 @@
 
 #pragma once
 
-#include <opc/ua/view.h>
-#include <opc/ua/attributes.h>
 #include <opc/ua/event.h>
-#include <opc/ua/node_management.h>
-#include <opc/ua/subscriptions.h>
+#include <opc/ua/services/attributes.h>
+#include <opc/ua/services/node_management.h>
+#include <opc/ua/services/view.h>
+#include <opc/ua/services/subscriptions.h>
 
 
 namespace OpcUa
