+++ include/opc/ua/services/endpoints.h
